backend/requirements-ci.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose==3.3.0
# Minimal ML dependencies for CI - using CPU-only versions
torch==2.0.1+cpu --find-links https://download.pytorch.org/whl/torch_stable.html
transformers==4.21.0
Pillow==10.0.0